+++ yandex/resource_yandex_kms_symmetric_key_iam_binding.go
@@ -1,0 +1,7 @@
+package yandex
+
+import "github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+
+func resourceYandexKMSSymmetricKeyIAMBinding() *schema.Resource {
+	return resourceIamBindingWithImport(IamKMSSymmetricKeySchema, newKMSSymmetricKeyIamUpdater, kmsSymmetricKeyIDParseFunc)
+}